Lily server process:   * Add an indexer kauri module and make it listen to the repository queue.   * Make hardcoded blobfs path (hdfs://locahost:9000) configurable.   * Some fiddling with dependencies to make everything work in Kauri:      - exclude Avro dependency from HBase, otherwise it is in Kauri's shared classloader, and it won't find the repository implementation classes. Seems like this needs to be improved in Avro: it should use the ContextClassLoader for loading classes instead of Class.forName.      - commons-httpclient needs commons-coded, but there were different version of commons-codec in use, causing it to be in the module-private classloaders, while commons-httpclient was being shared
Remote communication: use the HTTP connector for Avro, instead of the simple socket one. With the socketserver, I noticed that after the first avro request, the cpu of my laptop went to 190% and stayed like this until the server process was stopped. With upcoming avro 1.4 there will be a Netty-based connector that we could use.

Added a simple import tool to load field types, record types and records from a json file. The trigger to write this was to be able to test the standalone Lily without writing code. It is not yet feature complete (e.g. mixins not supported for record types, no possibility to update records) and there are no tests yet. Thanks to the power of Jackson (the json parser, not the singer) it was easy to do the parsing of the json in a combination of streaming and tree parsing, allowing any size of input json file to be processed.

Misc:
 * Remove non-existing artifacts from root pom.
 * Add missing AvroFieldTypeNotFoundException for getFieldTypeByName
 * Fixed switched repo and typemgr hosts in Client.


git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/projects/lily/trunk@4114 b80d95f8-e103-0410-bd55-e03a9dc29394