New 'Tester' tool: this is a tool for performing simple load testing. It runs CRUD operations in a loop, logs timings and errors, and runs until certain conditions are reached: maximum running time, maximum failures, maximum records created. It keeps all created records in memory to be able to verify the results of reading and for when performing updates. It runs only one thread, multiple processes can be started to perform concurrent testing. The schema of the created test records is configurable. The data is currently random generated, string fields are filled with a variable number of lorem ipsum words.
Import tool: split off the field/record type import logic into a separate class, and made it json-independent. This is generally useful utility code: it does things like checking if the field/record-type already exists, if so comparing if it is the same, updating it if necessary or reporting an error in case of conflict. Reused this in the Tester tool.


git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/projects/lily/trunk@4125 b80d95f8-e103-0410-bd55-e03a9dc29394