Adding 'books sample':  * sample import file  * matching sample indexer config  * matching SOLR schema.xml
Add special handling in import tool for vtag fields:
 * check they match the structural requirements (non-versioned, long type)
 * support special 'last' value for vtag fields, which will cause them to be automatically set to the last version. This part is currently tricky, and is done by doing a second update on the record. This is probably something which will need native support inside the repository to do it well.

For the import tool and the tester, removed the default arguments specified in the pom.xml so that they can be supplied via the command line. See updated README.txt for sample.

Use the standard zookeeper port (2181) in HBaseRunner and all sample configs.


git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/projects/lily/trunk@4131 b80d95f8-e103-0410-bd55-e03a9dc29394