RowLog subscriptions and listeners are configured through the RowLogConfigurationManager. (#49) MessageQueue instantiation is pulled outside of HBaseRepository. (#62) RowLogMessageListener classes are registered on the ListenerClassMapping.
Still open:
Indexer code to be updated to work with the RowLogConfigurationManager and ListenerClassMapping.
A RemoteListenerManager acts as an entrypoint to register remote listeners and wraps around the different RemoteListenerHandlers
Similar for a LocalListenerManager.
ListenerClassMapping now accepts one RowLogMessageListener class per subscriptionId. This should be refined to the level of Listeners.



git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/trunk@4297 b80d95f8-e103-0410-bd55-e03a9dc29394