In the RowLogShard only scan messages starting from a certain timestamp if its known that all older messages have been processed. (#167)
Refactored RowLogMessage to no longer contain a messageId but a creation timestamp instead. The message is uniquely identified by its rowkey, seqnr and timestamp. (#187)
The encodings on the global queue, row-local queue and in the communication to messageHandlers is now more efficient since it doesn't contain duplicate data.
This has an impact on backward compatibility. All existing messages must have been processed before upgrading to this code.



git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/trunk@4489 b80d95f8-e103-0410-bd55-e03a9dc29394