TypeManager ConcurrentUpdateTypeException fix: when you do createRecordType, and then if that fails (because of RecordTypeExistsExc) you fall back to updateRecordType, then this would always throw ConcurrentUpdateTypeException because the name reservation was not removed.
For field types this problem does not exist as the checkConcurrency method is called after checking if the field type already exists. Either way, IIUC, nothing guarantees that the cache is up to date (cfr. modifications happening on other nodes or threads), so the existence check seems unreliable to me.

Supply the fully qualified type name to ConcurrentUpdateTypeException.


git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/trunk@5060 b80d95f8-e103-0410-bd55-e03a9dc29394