Indexer: dynamicField/@continue attribute and FirstValueFormatter.
Added a new continue attribute on dynamicField which allows to change the default 'stop on first match' behavior of dynamic fields. One case where this is useful is to index the same field with a different formatter.

Added a FirstValueFormatter which behaves as the default formatter but only retains the first value in case of multi-valued fields.


git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/trunk@5069 b80d95f8-e103-0410-bd55-e03a9dc29394