getInputStream args, ValueType.isIndexBased, HierarchyPath extends from ArrayList
On Repository, re-introduced the getInputStream and getBlob methods with separate multi-value and hierarchy indexes. The previous approach was not really backwards compatible as it did not support null values (the indexer tests broke on it). One solution would have been to filter out the nulls, which I didn't like much, also, there was little reason to use Integer objects for the varargs, as this would imply null does something useful. The disadvantage is that Repository now has again 3 extra methods.

Introduced a method ValueType.isIndexBased() which returns true for list and path, both of which have values which are ordered lists.

Made HierarchyPath extend from ArrayList. This allows uniform handling of HierarchyPath and List, which are the same except for semantics. Technically not fully backwards compatible as previously you modify the array returned by getElements().


git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/branches/complextypes@5074 b80d95f8-e103-0410-bd55-e03a9dc29394