Adjust formatters for the changes related to complex field types.
DefaultFormatter:
 - adjusted for formatting recursive lists and records
 - now better extensible for customizing the behavior

Dropped the fallback-based mechanism of formatter selection, and the related configuration in the indexerconf. Formatters can now only be selected by name, and you can specify a custom default formatter. The reason is that the concepts of primitive type with multivalue/hierarchical qualifiers have been dropped/changed. This is of course backwards incompatible, the Formatter interface changed as well. As we only had one formatter implementation and no way to register custom ones, this should not pose a problem.

Once we have the ability to register custom ValueType implementations, we should review how you can register corresponding formatters for use in the indexer, especially as then you'll want the different formatters to play together to handle recursive values.


git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/branches/complextypes@5082 b80d95f8-e103-0410-bd55-e03a9dc29394