Indexer: adjusting dynamic fields to the new value type system. This change is partly backwards incompatible, existing indexerconf's using dynamic links will need to be reviewed.
The matching of types is more complex than before because types can now be arbitrarily nested. A type pattern syntax to match on types has been introduced. The use of <> in the type syntax is somewhat inconvenient in the indexerconf XML as it needs escaping.

The expressions to generate the dynammic field names become more difficult too. A few new variables have been introduced to insert the name of the nested type and deepest nested type.

Still todo: test case.


git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/branches/complextypes@5096 b80d95f8-e103-0410-bd55-e03a9dc29394