Make the Avro/Netty communication handle requests concurrently.
For a single, multi-threaded client talking to Lily, this can seriously multiply its performance.
The thing is that in Netty, the handling is done by a single thread, as it assums by default the processing is cpu-bound. Inserting an ExecutionHandler into the pipeline allows for long, io-bound, work to happen. Had to upgrade to a modified avro to be able to add this ExecutionHandler (https://issues.apache.org/jira/browse/AVRO-976).
Meanwhile upgraded Netty dependency.


git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/trunk@5254 b80d95f8-e103-0410-bd55-e03a9dc29394