MessagesWorkQueue: switch ArrayList to LinkedList.
Since most of the time we're removing elements from the start, and since we don't need random access to the list, a LinkedList will most likely be more efficient. Note that sometimes we skip messages (from rows for which we are already processing a message), otherwise we could have used a circular buffer.


git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/trunk@5272 b80d95f8-e103-0410-bd55-e03a9dc29394