Let Avro generate Java classes that use String instead of CharSequence.
This is a new option to the Avro schema compiler (and maven plugin) in Avro 1.6. This makes things easier and more efficient for us, as we needed strings anyway (and maps were difficult with charsequence keys).
Meanwhile removed the paranamer maven plugin, iirc that was once needed by avro but not anymore.


git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/trunk@5276 b80d95f8-e103-0410-bd55-e03a9dc29394