launch-test-lily: add ability to create a template data directory to be used for future startups.
This improves startup speed. The way it works is that when invoked with --prepare, it starts all services, stops them, and copies the data directory to ~/.lily/launcher/template/{version}. On subsequent startups, if this directory exists, it will be automatically used as basis for the new temporary data directory.


git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/branches/BRANCH_1_1_X@5314 b80d95f8-e103-0410-bd55-e03a9dc29394