Extending the HBase interruptability patch.
I noticed frequent hangs when stopping LilyClient (more specifically the CacheRefresher thread) when running tests with Lily embed, but this could equally well occur in any other case where we want to interrupt an HBase thread.
Additionally, when running LilyClient programs, you often see an ugly stacktrace at the end because of the interruption of the CacheRefresher thread, which we can't detect because HBase didn't set the original IE as cause.
Meanwhile extended the hbase fork instructions with usage of git to keep track of changes (more practical since then your IDE is aware of the changes).


git-svn-id: svn+ssh://dev.outerthought.org/outerthought/lilyproject/trunk@5516 b80d95f8-e103-0410-bd55-e03a9dc29394