Initial changes to make Lily compile and run with CDH4.
Still lots of things to fix and test, but at least it compiles and the base testcases run.
Synced forked classes (except HBase-MR fork as that will be dropped anyway).
Since HBase has no direct dependency on Hadoop anymore (they are included through a profile), I added those deps to lily-hbase-client. The HBaseExclusionsMojo was adjusted to handle this new situation.