LILY-364 Introducing smart population of the MQ so that each indexer only receives events about relevant records.
This makes it more efficient to have multiple indexes, especially when each index handles a different subset of the records.

To do this more selective MQ feeding, MessageQueueFeeder is replaced by a new IndexAwareMessageQueueFeeder.

In case the information based on which a record is included in some index changes, the old index(es) of which the record was part still need to receive the event in order to delete the record from their index. Thus, IndexAwareMessageQueueFeeder needs both the old&new values of this information, and this is made available in the RecordEvent by the IndexSelectionRecordUpdateHook, which plugs in in the critical path of the CRUD operations.

There is a new Kauri module called indexer-integration which registers the IndexAwareMessageQueueFeeder and IndexSelectionRecordUpdateHook (which needs to be done before the Repository starts).

In the indexerconf, the simple List<IndexCase> is replaced by IndexRecordFilter. The indexerconf.xml <records> section syntax changed (it is now replaced by a <recordFilter>), but the old syntax stays supported.

All this is still work in progress: the major components are in place and there is a basic testcase, but lots of things left to do.