testfw: add ability to wait for index to be known to IndexerRegistry. This is useful for testcases that make use of synchronous indexing.
This is implemented similar as for the rowlog subscription check: the IndexerRegistry exposes via JMX the list of indexes it knows about.

Meanwhile, changed behavior of the methods that have wait flags to throw an exception if the timeout passed. Previously, a boolean was returned, but in many cases this boolean was not checked. This is a backwards incompatible change.

Dropped the methods to add an index from file (because I have not encountered any use of it, and the number of methods to add indexes was exploding), made the method to add an index from byte[] public.