IndexUpdater fix: when a record changes so that a different <include> rule from the indexerconf matches, a different set of vtags might become active, in such case deletes must be done on solr to remove the entries corresponding to the old vtags.
For the IndexUpdater to be able to do this, it needs to know what <include> rule matched before, hence it needs both the old & new record state. This is exactly the same as what we need for evaluating the IndexRecordFilter in the IndexAwareMQFeeder, and is put into RecordEvent by IndexSelectionRecordUpdateHook. Therefore, the information provided by this hook is now also useful for IndexUpdater.

Additionally, for update of denormalized data to work, records which are not indexed by themselves but are accessed through dereference expressions also need to be matched by the IndexRecordFilter. To be able to express this, we need to support include rules with empty vtags: <include {conditions} vtags=""/>. This basically says: include the record, but don't index it. The same could be achieved by <match> rules to exclude the record from indexing, but that's more expensive (the recordfilter decision can be made without even loading the record) and (currently) doesn't help much with dynamic fields either.

Solving all this brought a couple of other changes with it:

* Refactored VTaggedRecord: moved out things which only dependend on RecordEvent (and not on a loaded Record) into RecordEventHelper. Probably could use further refactoring to completely split the two because now VTaggedRecord is also used in cases where there is no RecordEvent. (and then maybe reintroduce a third party to combine them).

* indexer-integration dropped: merged it with engine due to circular dependencies for their tests (engine test now needed the IndexRecordFilterHook). Could have been solved also by splitting indexer-integration into two different projects, but decided that it's ok to have it together with engine.

* Renamed RecordEvent.IndexSelection to RecordEvent.indexRecordFilterData, renamed IndexSelectionRecordUpdateHook to IndexRecordFilterHook.

* Fixed NPE in IndexRecordFilterHook in case of a record update where no record type is specified.

* Pulled out the code to create Record's from the information stored in RecordEvent into a new utility class (IndexRecordFilterUtil) so that it can be reused by IndexUpdater.

* IndexerTest now needs an IndexRecordFilterHook, which depends on IndexerModel which was previously not used in this test.