Honor tables.xml settings when creating deref tables.
Both in case of IndexerWorker or batch index build.
Also, don't create deref tables when not necessary in case of batch index build.
Note that deref table names depend on index name. As such, it's impossible to have default entries in tables.xml, which also makes that CF-level settings such as compression are not applied. Will look into this tomorrow.