Fix wrongfully closed HBase connections in LilyClient.
LilyClient closes HBase connections which might still be in use by other processes, this is e.g. the case when using LilyClient from within lily-server, which is the case when launching a batch index build. To some extent this is due to the recently added HConnectionManager.deleteStaleConnection, which we added because apparently the ref-count of the connection was not maintained correctly (leading to connections not being closed because the system thinks they're still in use, something which would best be fixed). In addition, HBase 0.92 changed so that it no longer looks at the identity of the Configuration object to manage the connection, but rather at the connection related properties within that object. To make sure a Configuration object has its own, separate connection that we can safely dispose of, we can set the property HConstants.HBASE_CLIENT_INSTANCE_ID in the hbase conf.

This fixes the failing LilyProxyTest.