Fixing a maven circular dependency.
The recently introduced testcase in the indexermaster uses the lily-server-test-fw, which depends on almost everything (including all the indexing stuff), thus obviously causes a circular depenendency. Fixed by moving the test in a separate project. An alternative would be to not base the test on LilyProxy, but to just instantiate the parts of Lily we need (see RepositorySetup).