LILY-568 Fix the lib dir of the binary dist
Before we were using dependencies of the 'dist' module to gather artifacts and their runtime dependencies (in conjunction with the mvn assembly plugin).
Now we're using our own plugin to assemble artifacts and their transitive runtime dependencies. (The plugin is based on the library that's powering maven 3 - aether).

Some side notes:
- removed solr-core dependency - apparently this is no longer needed.
- The maven version bump required adding the maven-compat project here and there.