Removing locking in the repository write paths.
And removed the global/rowlock project.

The locking was based on the class RowLocker, which locked and unlocked by means of a checkAndPut call on HBase that wrote a timestamp (for lock expiry) combined with some random data to avoid lock stealing.

Instead, we now use optimistic locking. We can do this because we don't have the rowlog and execution of the WAL secondary actions in the write path anymore, which were the most important reason to have locking in there.

Introduced a new system field called occ ('optimistic concurrency control'). Basically a kind of version counter, but named occ because we already had a field called version with different semantics.

Number of ops for a record create is now 1 get + 1 checkAndPut. From here on, it will only get slower again (except for optimizing the scope of the get and put, and for having create-only records).

Fine points:
 * decorators & updatehooks are not executed within the lock anymore, so these should be careful with their side-effects
 * blob-related code is not executed within locks anymore, afaics behavior should stay correct
 * no RecordLockedException anymore, instead ConcurrentRecordUpdateException. BalancingAndRetryingLilyConnection has been adjusted in this regard.
 * the clearData() method called during record deletion (and also in case of record reincarnation, to be sure) is not executed within a lock anymore, so can potentially run concurrently with record re-creation. The method has therefore been adjusted to only treat the old versions. For the system fields I changed the deleteColumn calls to deleteColumns calls, since the former only deletes the latest version which was a bug.