Write payload & enable replication.
Write the recordevent MQ payload to a system column in the data cf. Goal is that the payload would become part of the waledit in the hlog, so that it is available to things reading the hlog. Currently only for create/update/update-mutable, not for deletes, will figure out later how we can get that in the waledit for that case.

Enabled replication (scope=1) for the data cf of the record table.