Introducing HBase-replication feeded indexing.
Adding sep (side effect processor -- name might change). This is an experiment to replace the rowlog MQ by instead listening to the HBase replication feed. This commit includes the minimal changes to make this work:

 * IndexerAdmin will add/remove replication peers to HBase when an index is added/removed (SepModel)
 * IndexerWorker will start a "fake hregionserver" that consumes the replication feed and delivers messages to the indexer (SepEventSlave)
 * It is assumed the payload is written under a certain key, currently only waledits which contain such a payload are forwarded to the listeners
 * Functionally missing is at least handling of delete events and creation of re-index events for denormalized data

It is recommended to configure your hbase to set replication.source.ratio to 1 for smaller clusters and to set replication.source.nb.capacity to for example 200 to avoid very large batches (default is 25000). For launch-test-lily 200 is the default.