LILY-912 linkindex via sep
Trigger the updating of the link index via the SEP instead of as a rowlog secondary action.

Incompatible changes:
 * the linkindex is now fed asynchronously
 * the flag to enable the linkindex is now in general/linkindex.xml.

Meanwhile introduced an interface for SepModel (needed the interface to be able to wire in Kauri) and moved the sep interfaces to lily-sep-api.

SepModel: added variants of add/removeSubscription that don't throw an exception if the subscription does or does not exist.