LILY-906 LILY-912 Migrate tests to new waitForSep
Had quite some difficulty getting the IndexerTest to work with the new waitForSep. The special thing about the IndexerTests is that it regularly removes and then re-adds an index with the same name (but different conf). The problems are the async nature of ReplicationSource start/stop and the fact that it doesn't unregister its MBean (which is used by waitForSep to read the log queue size), so that wrong/conflicting information is published in jmx. Added workarounds for this: the mbean of the replicationsource is unregistered manually by us, and we wait for replicationsources to be started/stopped by looking at the running threads (cfr. HBaseProxyMBean). Also made it so that a different index name is used each time.

There were some other things that took some time to figure out: IndexerTest.changeIndexUpdater sometimes fails on purpose (if the indexerconf is syntactically wrong) in which case the necessary sep cleanup of the previous index wasn't performed, and the waitForSep actually needs to be performed twice: once to process all current events, and once to process the reindex messages that might have been generated in the previous step.

Moved waitForSep to a more generic HBaseProxy.waitForReplication, so that it can also be used in the context of RepositorySetup. Also, before rolling the log, write a dummy change to be sure there is something in the hlog and thus that it will actually roll.