Adding new conf/general/hadoop.xml config file
This file is intended for general hadoop configuration, i.e. hdfs settings, and is inherited by both hbase.xml and mapreduce.xml. The fs.defaultFS moved by default from mapreduce.xml to hadoop.xml.

This is because the need has come up a few times to connect from within lily-server extensions to HDFS, in which case it was conceptually strange to rely on the mapreduce properties.