Renamed mixins to supertypes
For backwards compatibility, the old mixin-methods are kept on the Java API but have been deprecated. The REST API still accepts mixins on input, but always returns supertypes on output, this is the only API incompatibility. The mixin system fields (for use in the indexer) are also still supported but log deprecation warnings.