Add refreshSubtypes functionality
refreshSubtypes is a flag which can be passed to the updateRecordTypes method. If true, then when updating a supertype record type, its subtypes will be automatically updated to point to this new version of the supertype (pointers are always to specific versions of supertypes). This is a convenience feature, because otherwise it is kind of hard for the user to move all these versioned pointers.

In core API, builder API, avro. REST itf still todo.

Also used consistent casing for subtypes (not subTypes, following wikipedia).