FieldValueFltr support metadata + drop hbase-ext
Let FieldValueFilter also work for fields that have metadata (for which the cell format is different). This involves adding a custom HBase filter.

Moved hbase-ext code in with repository-impl. This wasn't very necessary right now (just reusing FieldFlags class), but will be necessary anyway once we have filters supporting metadata.

Impact is that hadoop-test-fw cannot have a dependency on hbase-ext, now repository-impl, anymore since this would result in circular dependencies. Decided to solve this by making a separate project, hadoop-test-fw-standalone, that does have dependencies on the things that Lily requires to be on the regionserver classpath (and removed these deps from hadoop-test-fw, which is not an issue for embed-mode since the stuff that needs to be on the classpath is usually also a maven dependency).

The launch-hadoop script is now in global/hadoop-test-fw-standalone/target