Runtime: extra logging + version cmp logic
Extended the artifact version comparison login of the runtime to support the style of suffixes we use for git (it can only compare them if the dates are different), and also support "-SNAPSHOT" (considered to be more recent than anything else, except than the same version without suffix, as that is assumed to be the final release).

Added more logging in case version numbers can't be compared.