waitOnReplication: don't use record table
but instead -ROOT- table. This makes that custom SEP listeners don't get confused by the trigger row we write.
The actual table name didn't actually matter (it was made an argument for cases where the 'record' table would not exist, but the -ROOT- table always exists), so removed any specific uses of that.