TableManager adaptations for multitenancy
RepositoryTableManager should now be per-tenant, and Repository serves as a factory for obtaining the tenant-specific RepositoryTableManager.

RepositoryTableManager did not yet have a remote equivalent, and since it now needs one, added the remote impl. I wrote all this code without any tests for it, so it probably doesn't work. To be done later.

I had to slightly refactor TableConfigBuilder as well as part of this, cfr the parsing of the input parameters into "byte[][] splitKeys".