Adjust importlib to be repository based
since the role of RepositoryManager is now to give tenant-specific repositories, and since most code should be tenant-scoped it needs to be written against Repository.

Also changed lily-scan-records to be based on LTable instead of Repository, just to check if the mulitenancy-related API changes make sense. Found a need to introduce a PureRepository to protect against unintended use of LTable methods through Repository.