core repo api change: introduce LRepository itf
Introduce an LRepository interface to which the methods from Repository, which were not yet moved to LTable, are moved. Repository is now just the combination of LRepository and LTable. Rationale for this is that otherwise using Repository will be error-prone, since it allows for unconsiously calling LTable methods on it. Once everything is moved to LRepository/LTable, we'll also have a better separation of concerns.

Also changed the signature of RepositoryManager.getRepository() to return LRepository objects. This actually had the widest impact, since in most places it required casting back to Repository.