Fix failing tests in connect mode
LilyProxyTest & KeepDataTest were failing in connect mode, essentially because they were using the same index name, and because the cleanup of replication peers didn't wait for them to be really removed, confusing the code which waits for things to be done. Using different index names in both tests would have solved it as well, but this is more solid.