import tool: add support for json-per-line format
Add support for a new input file format to the json import tool, whereby each line contains a full json object describing a single record. The json format of the record is the same as always. Namespace declarations need to be made per record, or you need to use fully qualified names. A schema cannot be included in this format.