if field value is invalid, give context in error
when supplying a wrong type of value for a field, e.g. when trying to store a long in a string, it would now just give a classcastexception within the valuetype serialization code without giving any clue as to what field contains the problematic value. Now providing more context and also added explicit check on value type.