json import: support mode also for record types
Support "mode: create/update/create_or_update" and the failIfExists & failIfNotExists settings also for record types (previously these existed only for records).

For example, "mode: create" is useful if you have a schema that declares a record type which is later on modified programmatically: reloading the schema from json will then not touch the record type if it already exists.