scan speedup when older record types are involved
LILY-2348
Reads or scans over records that point to older record types were significantly slower (> 10 times) because only the latest version of record types is cached.
There was in fact a simple solution, since the only reason a record type is retrieved while reading a record is to get the name of the record type, which is the same regardless the version of the record type.

Note that a similar problem still exists for nested records, where such an easy fix is not possible, there we'll really have to improve the schema cache.