Enable pass-through of authz ctx to hbase (1/x)
The idea behind this and follow up commits is to enable Lily to pass on an authorization context to a hbase-based authorization framework by NGDATA (not part of the OSS Lily DR).

 * define a thread-local based authorization context holder
 * wrap tables in AuthEnabledHTable to make sure authz context is passed on to hbase for each request
 * testcase

introduces a dependency on lily-hbase-authz-public-client which is not yet publicly available but will be soon.

not yet / todo: make the same work for the remoterepository, i.o.w. modify rpc protocol to pass on authz context