LILY-2537 pass on authz context over avro rpc
Pass on the authorization context from lily-client to lily-server when using the avro rpc.
Only added this for the methods that need it (thus for methods working on records, not for the typemgr methods).
There is no authentication on the avro rpc, we are just passing on trusted information.
Noticed that there is a read-only method (getVariants) that goes over rpc, this should not be the case.