LILY-2548 LILY-2537 remove use of AuthorizationContextProvider
Removed the ability to provide your own AuthorizationContextProvider to LilyClient or RepositoryManager, instead use the fixed DRAuthorizationContextProvider as input to AuthEnabledHTable.
Decided this flexibility was not needed, besides there could be different sources (authz context coming from rpc or rest). So it was simpler to have a model where the authz context is always set into the same threadlocal variable.